Move Predictive handlers to observational module (#550)
* Move cut posterior handlers to robust module

* moving nmc and predictive functionality from robust to observational

* format and lint

* docstring ref

* restore nmc

* inline bind_leftmost_dim in nmc

* format

* revert format

* revert format

* Update utils.py

---------
