+++ chirho/indexed/handlers.py
@@ -4,6 +4,7 @@ from typing import Any, Callable, Dict, Hashable, List, Optional
 
 import pyro
 import torch
+from typing_extensions import ParamSpec
 
 from chirho.indexed.internals import (
     _LazyPlateMessenger,
@@ -11,6 +12,8 @@ from chirho.indexed.internals import (
     get_sample_msg_device,
 )
 from chirho.indexed.ops import union
+
+P = ParamSpec("P")
 
 
 class IndexPlatesMessenger(pyro.poutine.messenger.Messenger):
@@ -137,3 +140,24 @@ class DependentMaskMessenger(pyro.poutine.messenger.Messenger):
         msg["fn"] = msg["fn"].expand(
             torch.broadcast_shapes(msg["fn"].batch_shape, mask.shape)
         )
+
+
+@pyro.poutine.block()
+@pyro.validation_enabled(False)
+@torch.no_grad()
+def guess_max_plate_nesting(
+    model: Callable[P, Any], guide: Callable[P, Any], *args: P.args, **kwargs: P.kwargs
+) -> int:
+    """
+    Guesses the maximum plate nesting level by running `pyro.infer.Trace_ELBO`
+
+    :param model: Python callable containing Pyro primitives.
+    :type model: Callable[P, Any]
+    :param guide: Python callable containing Pyro primitives.
+    :type guide: Callable[P, Any]
+    :return: maximum plate nesting level
+    :rtype: int
+    """
+    elbo = pyro.infer.Trace_ELBO()
+    elbo._guess_max_plate_nesting(model, guide, args, kwargs)
+    return elbo.max_plate_nesting
