+++ tests/robust/test_handlers/test_estimators.py
@@ -6,11 +6,14 @@ import pytest
 import torch
 from typing_extensions import ParamSpec
 
+from chirho.observational.handlers.predictive import (
+    PredictiveFunctional,
+    PredictiveModel,
+)
 from chirho.robust.handlers.estimators import (
     MonteCarloInfluenceEstimator,
     one_step_corrected_estimator,
 )
-from chirho.robust.handlers.predictive import PredictiveFunctional, PredictiveModel
 
 from ..robust_fixtures import SimpleGuide, SimpleModel
