+++ tests/robust/test_performance.py
@@ -9,12 +9,12 @@ import pytest
 import torch
 from typing_extensions import ParamSpec
 
-from chirho.indexed.handlers import DependentMaskMessenger
+from chirho.indexed.handlers import DependentMaskMessenger, guess_max_plate_nesting
 from chirho.observational.handlers import condition
-from chirho.robust.handlers.predictive import PredictiveModel
+from chirho.observational.handlers.predictive import PredictiveModel
 from chirho.robust.internals.linearize import make_empirical_fisher_vp
 from chirho.robust.internals.nmc import BatchedNMCLogMarginalLikelihood
-from chirho.robust.internals.utils import guess_max_plate_nesting, make_functional_call
+from chirho.robust.internals.utils import make_functional_call
 from chirho.robust.ops import Point
 
 from .robust_fixtures import SimpleGuide, SimpleModel
